+++ src/utils/filters.js
@@ -1,24 +1,36 @@
-import { CAPABILITIES, EDIT_CAPABILITIES_OPTIONS } from '../constants';
+import {
+  CAPABILITIES,
+  CRITERIA,
+  EDIT_CAPABILITIES_OPTIONS,
+} from '../constants';
 
-export const isCapabilityDisabled = (capabilityValue, perms = {}) => {
-  const { isUserRadioDisabled, isInventoryRadioDisabled } = perms;
+export const isCapabilityDisabled = (capabilityValue, view, perms = {}) => {
+  const isQuery = view === CRITERIA.QUERY;
+
+  const {
+    isUserRadioDisabled,
+    isInventoryRadioDisabled,
+    hasItemsPerms,
+    hasHoldingsPerms,
+    hasUsersPerms,
+  } = perms;
 
   const capabilitiesMap = {
-    [CAPABILITIES.USER]: isUserRadioDisabled,
-    [CAPABILITIES.ITEM]: isInventoryRadioDisabled,
-    [CAPABILITIES.HOLDING]: isInventoryRadioDisabled,
+    [CAPABILITIES.USER]: isQuery ? !hasUsersPerms : isUserRadioDisabled,
+    [CAPABILITIES.ITEM]: isQuery ? !hasItemsPerms : isInventoryRadioDisabled,
+    [CAPABILITIES.HOLDING]: isQuery ? !hasHoldingsPerms : isInventoryRadioDisabled,
   };
 
   return capabilitiesMap[capabilityValue];
 };
 
-export const getCapabilityOptions = (perms) => EDIT_CAPABILITIES_OPTIONS.map(capability => ({
+export const getCapabilityOptions = (view, perms) => EDIT_CAPABILITIES_OPTIONS.map(capability => ({
   ...capability,
-  disabled: isCapabilityDisabled(capability.value, perms),
+  disabled: isCapabilityDisabled(capability.value, view, perms),
 }));
 
-export const getDefaultCapabilities = (perms) => {
-  const capabilityOptions = getCapabilityOptions(perms);
+export const getDefaultCapabilities = (view, perms) => {
+  const capabilityOptions = getCapabilityOptions(view, perms);
 
   return capabilityOptions.find(option => !option.disabled)?.value;
 };